done downloading

+++ data/views-download.sh
@@ -2,5 +2,5 @@
 
 mkdir -p downloads/views
 for viewid in $(./datasets-parse.py); do
-  wget -O "downloads/views/${viewid}.json" "https://data.cityofnewyork.us/views/${viewid}.json"
+  test -e "downloads/views/${viewid}.json" || wget -O "downloads/views/${viewid}.json" "https://data.cityofnewyork.us/views/${viewid}.json"
 done
